Bug fix: Overwriting PAREN_CLOSE is bad as it will leave the level hanging and thus break a lot of things like indentation etc.
The whole concept of marking CT_OC_CLASS_EXT is flawed as an extension differs from a category 
in that it is empty parens with no word in between. To mark this without overwriting one parens 
type we could only flag the parent type as OC_CLASS_EXT but by doing so we might break other stuff 
depending on when we are in an OC_IMPL or OC_INTF which would be the parent type otherwise. 
In fact, I am pretty sure we would break other things. 

Marking OC_CLASS_EXT isn't good for much anyway so I will just remove it completely for now.