Enhancement: ObjC colon alignment now has quite a few new tricks up its sleeve.
The alignment system for ObjC msg sends has been completely reworked as it became apparent
that a lot more edge cases need to be handled. Take a look at the new accompanying test 
(colon_align.m) to get a taste.

- It now correctly handles nested msg sends passed as args within the original bracket level.
- It now correctly handles commented-out lines in between multi-line spaced sends.
- It now has the ability to switch behavior based on a new settings when dealing with msg send
  lines which have multiple colons/args belonging to the same bracket level.
  The new option allows to choose the first or the last of the colons as alignment point.
- It msg sends alone which appear to be multi-line spaced but are in fact just a single line with
  args spaced over multiple lines as is common with aligning multiple string literals where each literal
  gets its own line or dictionary keys, etc.

In principle this works by "noting" the position of the colon on the first line which then serves as the 
"fix point". Otherwise the first line is left alone. The action starts on following line. The idea is to
add the first chunk of each line to the align stack and calculate a column adjustment with regards to 
the fix point. To get this adjustment the delta between each line's reference colon and the fix point is
calculated. The reference colon is most often the only colon on the line or depending on settings either
the first or the last. This delta is then used in a special way in align_to_column during AlignStack::Flush(). 
There's a multitude of fail-out conditions and edge-case stuff along the way.

Granted, a fair amount of code has been introduced, but a good portion of that is convenience functions
which may serve other purposes in the future.